gRPC-server/requirements.txt
kafka-python==2.0.2
elasticsearch==7.17.4
numpy==1.19.5
scipy
scikit-learn==0.24.2
pandas==1.1.5
grpcio-tools==1.47.0
grpcio-reflection==1.47.0